refactor: extract text editor state hook
